ConfigurationForModels&Mirgation

+++ WalletTracker/Models/SavingsGoals.cs
@@ -4,8 +4,9 @@ public class SavingsGoals
 {
     public Guid Id { get; set; }
     public string? Name { get; set; }
-    public double Target { get; set; }
-    public double Current { get; set; }
+    public decimal Target { get; set; }
+    public decimal Current { get; set; }
+    
     public DateTime TargetDate { get; set; }
     
     public Guid UserId { get; set; }
